requirements.txt
pillow
piexif
pyyaml
ppretty
requests
